### 2025-04-20 #8082 @danielskatz
👋 @ZhiweiWei-NAMI - Thanks for your submission.

While I look for an editor, I have a question and a request:

Question: In the repository, I see a recent commit "[adding pypi publish workflow](https://github.com/ZhiweiWei-NAMI/AirFogSim/commit/e82e7250ddb0c5eeee5693658fec7a278633eb1b)" but I don't see PyPI mentioned in the README. Are you planning to update the README?

Request: Please work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
