### 2025-04-21 #8082 @ZhiweiWei-NAMI
> 👋 [@ZhiweiWei-NAMI](https://github.com/ZhiweiWei-NAMI) - Thanks for your submission.
>
> While I look for an editor, I have a question and a request:
>
> Question: In the repository, I see a recent commit "[adding pypi publish workflow](https://github.com/ZhiweiWei-NAMI/AirFogSim/commit/e82e7250ddb0c5eeee5693658fec7a278633eb1b)" but I don't see PyPI mentioned in the README. Are you planning to update the README?
>
> Request: Please work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Hi @danielskatz,

Thank you for your feedback! I have made the following updates based on your request:

1. **PyPI Installation in README**: I have updated the README to include installation instructions from PyPI. The new section added is:

   #### Option 1: Install from PyPI (Recommended)

   The easiest way to install AirFogSim is directly from PyPI:

   ```bash
   pip install airfogsim
   ```
2. DOI for OMNeT++ and SUMO: I have addressed the missing DOI for the references:
- OMNeT++ Discrete Event Simulation System: Added a URL in place of DOI.
- Recent Development and Applications of SUMO: Also added a URL for this reference.
I have updated the .bib file accordingly and generated the new PDF.

Thank you!
