### 2025-05-06 #8082 @fxjung
> Hi [@fxjung](https://github.com/fxjung) do you have time to review the paper?

@diehlpk Thanks for the invitation! Unfortunately I'm very busy at the moment and wouldn't be able to produce a timely review.
