### 2025-05-13 #8082 @SongshGeo
@diehlpk Thanks for the invitation. However, I found my abilities may not be suitable for reviewing the code. I don't have enough knowledge about 3d or drones. Learning the relevant code might require a significant time investment, which doesn't align with my current schedule. :) Anyhow, it seems like a very interesting project!
