### 2025-05-13 #8082 @jmlarson1
I must decline. I do not have experience with
- discrete-event simulation (of autonomous agents)
- SimPy
- UAV-integrated computing environments
- modeling interactions between aerial/terrestrial nodes
to review this.
