### 2025-05-25 #8267 @editorialbot
[REVIEW]: AirFogSim: A Python Package for Benchmarking Collaborative Intelligence in Low-Altitude Vehicular Fog Computing
**Submitting author:** @ZhiweiWei-NAMI (<a href="http://orcid.org/0000-0001-9593-0013">Zhiwei Wei</a>)
**Repository:** https://github.com/ZhiweiWei-NAMI/AirFogSim
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.0
