### 2025-06-05 #8267 @ZhiweiWei-NAMI
> Hi everyone, I've had a good look through the AirFogSim paper and the repository, and I think this is a really interesting and potentially very useful package for researchers in the Low-Altitude Vehicular Fog Computing space! The motivation is clear, and tackling these complex simulations with a Python-based, agent-based approach is a great direction. The integration with tools like SUMO and the built-in visualization are also strong points. 👏
>
> To align AirFogSim more closely with JOSS standards and make it even more impactful for the community, there are a couple of primary areas that I think would need some focus:
>
> 1. **Automated Tests:** This is a cornerstone for JOSS submissions. Adding a suite of automated tests to the repository would be essential to ensure the software's reliability, catch regressions, and make it easier for others to contribute confidently.
> 2. **Comprehensive Documentation (in the repo):**
>
>    * **Installation:** Clear, step-by-step instructions, including all dependencies (Python version, FastAPI, React, etc.), ideally managed with a solution like `requirements.txt` or an environment file.
>    * **Community Guidelines:** As per JOSS guidelines, clear instructions on how to contribute, report issues, and seek support would be very helpful. A `CONTRIBUTING.md` file is often a good place for this.
>
> The groundwork here is really solid. Addressing the automated testing and expanding the repository documentation would be the key next steps. I'm looking forward to seeing how this develops!

Thank you so much for your thorough review and valuable feedback! We are making progress on these fronts and will commit to addressing these key areas.
