### 2025-06-16 #8267 @mengdayayyt
Thanks for the opportunity to review AirFogSim! This is impressive work, and I can see it becoming a valuable tool for the LAVFC community. I have a few suggestions that I believe would significantly enhance the package:

1.  **Elaborating on "Collaborative Intelligence" Benchmarking:** This is highlighted as a key feature. It would be fantastic if the documentation (or even a dedicated section/example in the paper if space permits) could more explicitly guide a user on *how* to set up simulations to benchmark different CI strategies. What specific metrics does AirFogSim provide or suggest for this? How can users easily compare outcomes of different collaborative approaches?
2.  **API Documentation:** Having API documentation for the core classes and functions (e.g., for `Agent`, `Component`, `Workflow`, `DataProvider`) would be immensely helpful for users who want to extend the simulator or understand its internals more deeply. Tools like Sphinx with `autodoc` can be great for this.
3.  **Advanced Examples:** Building on the existing example, a tutorial-style example that walks through a more complex scenario, perhaps one that really leverages the `DataProvider` for external data and demonstrates a non-trivial `Workflow`, would be very beneficial for new users.

The package has a strong foundation. Focusing on making the "collaborative intelligence" benchmarking more tangible for users through examples and metrics, alongside richer API documentation, would really elevate it. Great job so far, and I'm keen to see the next iteration!
