### 2025-06-30 #8267 @diehlpk
@ZhiweiWei-NAMI

- I would recommend to move Figure 2 to the documentation since the font is too small and it is hard to read.
