### 2025-07-01 #8267 @ZhiweiWei-NAMI
Hi @diehlpk,

Thank you very much for your review and the helpful suggestions.

I have now addressed all the points you raised. I moved Figure 2 to the documentation and have completed the pre-publication checklist below:

Here is the requested information for the new release:

* **Version:** `v1.1.1`
* **DOI:** `10.5281/zenodo.15779001`

Thank you again for your guidance through this process. Please let me know if anything else is needed.
