### 2025-07-01 #8267 @ZhiweiWei-NAMI
@diehlpk Thank you for catching that. I have removed the incorrect citation from `paper.md`.
