### 2025-07-01 #8267 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
