### 2025-07-01 #8267 @danielskatz
Congratulations to @ZhiweiWei-NAMI (Zhiwei Wei) and co-authors on your JOSS publication!!
