### 2025-08-19 #8267 @ZhiweiWei-NAMI
Hi @danielskatz ,

Thank you again for your work on our submission. We would like to request a change to the paper's title.

* **Current Title:** AirFogSim
* **Proposed New Title:** AirFogSim++

**Reason for the change:**
We feel the new title, "AirFogSim++", more accurately reflects the significant architectural and functional evolution of our software.

The name "AirFogSim" was associated with our previous work, which used a step-wise method to simulate specific tasks in UAV-assisted Vehicular Fog Computing (VFC) scenarios.

This submission, however, is a fundamentally new version built on agent-based modeling. This new architecture makes the simulator much more powerful and flexible, allowing it to support more general and complex scenarios beyond the initial scope. The "++" suffix is intended to clearly signify this major advancement and distinguish it from the prior, more limited version.

Please let us know if this change is acceptable. If approved, I will update the `paper.md` file and the package in our repository immediately.

Thank you for your consideration.
