### 2025-08-19 #8267 @danielskatz
@ZhiweiWei-NAMI - I'm sorry, but we will not change a published paper unless there is an error that appears in or immediately after the publication process.

This paper is part of the scholarly record at this point. You can certainly change the name of the software going forward, but the archive of the software and the paper are static at this point.
